pyproject.toml
[project]
name = "cnesdata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anyio==4.12.1",
    "backoff==2.2.1",
    "basedosdados==2.0.3",
    "boolean-py==5.0",
    "cachecontrol==0.14.4",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.6",
    "colorama==0.4.6",
    "cryptography==46.0.5",
    "cyclonedx-python-lib==11.7.0",
    "db-dtypes==1.5.0",
    "defusedxml==0.7.1",
    "et-xmlfile==2.0.0",
    "fastavro==1.12.1",
    "fdb==2.0.4",
    "faker>=30",
    "filelock==3.25.2",
    "firebird-base==2.0.2",
    "firebird-driver==2.0.2",
    "google-api-core==2.30.0",
    "google-api-python-client==2.193.0",
    "google-auth==2.49.1",
    "google-auth-httplib2==0.3.0",
    "google-auth-oauthlib==1.3.0",
    "google-cloud-bigquery==3.40.1",
    "google-cloud-bigquery-connection==1.20.0",
    "google-cloud-bigquery-storage==2.36.2",
    "google-cloud-core==2.5.0",
    "google-cloud-storage==3.10.0",
    "google-crc32c==1.8.0",
    "google-resumable-media==2.8.0",
    "googleapis-common-protos==1.73.0",
    "gql==4.0.0",
    "graphql-core==3.2.8",
    "greenlet==3.3.2",
    "grpc-google-iam-v1==0.14.3",
    "grpcio==1.78.0",
    "grpcio-status==1.78.0",
    "holidays>=0.46",
    "httplib2==0.31.2",
    "idna==3.11",
    "iniconfig==2.3.0",
    "license-expression==30.4.4",
    "loguru==0.7.3",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "msgpack==1.1.2",
    "multidict==6.7.1",
    "numpy==2.4.3",
    "oauthlib==3.3.1",
    "openpyxl==3.1.5",
    "packageurl-python==0.17.6",
    "packaging==26.0",
    "pandas==2.3.3",
    "pandas-gbq==0.34.0",
    "pandavro==1.9.0",
    "pandera==0.30.1",
    "pip-api==0.0.34",
    "pip-audit==2.10.0",
    "pip-requirements-parser==32.0.1",
    "platformdirs==4.9.4",
    "plotly>=5.20.0",
    "pluggy==1.6.0",
    "propcache==0.4.1",
    "proto-plus==1.27.1",
    "protobuf==6.33.6",
    "psycopg[binary]>=3.2",
    "psycopg2-binary>=2.9",
    "psutil==7.2.2",
    "py-serializable==2.1.0",
    "pyarrow==23.0.1",
    "pyasn1==0.6.3",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydata-google-auth==1.9.1",
    "pygments==2.19.2",
    "pyparsing==3.3.2",
    "alembic>=1.14",
    "pytest==9.0.2",
    "pytest-docker>=3.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "pytz==2026.1.post1",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.3.3",
    "ruff==0.15.7",
    "setuptools==82.0.1",
    "six==1.17.0",
    "sortedcontainers==2.4.0",
    "sqlalchemy==2.0.48",
    "sqlalchemy-firebird==2.1",
    "streamlit>=1.56.0",
    "streamlit-aggrid==1.2.1.post2",
    "tenacity==9.1.2",
    "tomli==2.4.0",
    "tomli-w==1.2.0",
    "tomlkit==0.14.0",
    "tqdm==4.67.3",
    "typing-extensions==4.15.0",
    "tzdata==2025.3",
    "uritemplate==4.2.0",
    "urllib3==2.6.3",
    "win32-setctime==1.2.0",
    "yarl==1.23.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "httpx>=0.28",
    "polars>=1.20",
    "asyncpg>=0.30",
]
[tool.uv]
   override-dependencies = ["protobuf==6.33.6"]

[tool.pyright]
# fdb is a legacy Firebird driver with no type stubs — suppress the false-positive "Cannot find module" error
reportMissingModuleSource = "none"
